requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.0 